refactor(bigquery): share test helpers with samples
