+++ src/bigquery/examples/src/query.rs
@@ -41,22 +41,11 @@ mod script;
 mod simple_app;
 mod total_rows;
 
+use crate::{create_dataset, delete_dataset, random_id_suffix};
 use google_cloud_bigquery_v2::client::DatasetService;
-use google_cloud_bigquery_v2::model::{Dataset, DatasetReference};
 use google_cloud_test_utils::runtime_config::project_id;
-use rand::{RngExt, distr::Alphanumeric};
 use std::future::Future;
 use std::pin::Pin;
-
-const INSTANCE_LABEL: &str = "rust-sdk-integration-test";
-
-fn random_id_suffix() -> String {
-    rand::rng()
-        .sample_iter(&Alphanumeric)
-        .take(8)
-        .map(char::from)
-        .collect()
-}
 
 pub async fn run_samples() -> anyhow::Result<()> {
     let project_id = project_id()?;
@@ -96,18 +85,7 @@ pub async fn run_samples_with_resources() -> anyhow::Result<()> {
     let project_id = project_id()?;
     let dataset_service = DatasetService::builder().build().await?;
     let dataset_id = format!("rust_bq_samples_{}", random_id_suffix());
-
-    println!("Creating sample dataset `{dataset_id}`...");
-    dataset_service
-        .insert_dataset()
-        .set_project_id(&project_id)
-        .set_dataset(
-            Dataset::new()
-                .set_dataset_reference(DatasetReference::new().set_dataset_id(&dataset_id))
-                .set_labels([(INSTANCE_LABEL, "true")]),
-        )
-        .send()
-        .await?;
+    let _ = create_dataset(&dataset_service, &project_id, &dataset_id).await?;
 
     let destination_table_id = format!("dest_{}", random_id_suffix());
     let dest_legacy_table_id = format!("dest_legacy_{}", random_id_suffix());
@@ -150,13 +128,7 @@ pub async fn run_samples_with_resources() -> anyhow::Result<()> {
         .collect();
 
     println!("Deleting sample dataset `{dataset_id}`...");
-    let _ = dataset_service
-        .delete_dataset()
-        .set_project_id(&project_id)
-        .set_dataset_id(&dataset_id)
-        .set_delete_contents(true)
-        .send()
-        .await;
+    delete_dataset(&dataset_service, &project_id, &dataset_id).await?;
 
     let _ = res?;
     Ok(())
