+++ tests/bigquery/src/dataset.rs
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::{INSTANCE_LABEL, random_id_suffix};
 use anyhow::Result;
-use futures::stream::{StreamExt, TryStreamExt};
+use bigquery_samples::{
+    INSTANCE_LABEL, cleanup_stale_datasets, create_dataset, delete_dataset, random_dataset_id,
+};
+use futures::stream::TryStreamExt;
 use google_cloud_bigquery_v2::client::DatasetService;
-use google_cloud_bigquery_v2::model::{Dataset, DatasetReference};
-use google_cloud_gax::{error::rpc::Code, paginator::ItemPaginator};
+use google_cloud_gax::paginator::ItemPaginator;
 use google_cloud_test_utils::runtime_config::project_id;
 
 pub async fn dataset_admin() -> Result<()> {
@@ -46,130 +47,4 @@ pub async fn dataset_admin() -> Result<()> {
     println!("DELETE DATASET");
 
     Ok(())
-}
-
-pub(crate) async fn create_dataset(
-    client: &DatasetService,
-    project_id: &str,
-    dataset_id: &str,
-) -> Result<Dataset> {
-    println!("CREATING DATASET WITH ID: {dataset_id}");
-    let ds = client
-        .insert_dataset()
-        .set_project_id(project_id)
-        .set_dataset(
-            Dataset::new()
-                .set_dataset_reference(DatasetReference::new().set_dataset_id(dataset_id))
-                .set_labels([(INSTANCE_LABEL, "true")]),
-        )
-        .send()
-        .await?;
-    Ok(ds)
-}
-
-pub(crate) async fn delete_dataset(
-    client: &DatasetService,
-    project_id: &str,
-    dataset_id: &str,
-) -> Result<()> {
-    client
-        .delete_dataset()
-        .set_project_id(project_id)
-        .set_dataset_id(dataset_id)
-        .set_delete_contents(true)
-        .send()
-        .await?;
-    Ok(())
-}
-
-pub(crate) async fn cleanup_stale_datasets(
-    client: &DatasetService,
-    project_id: &str,
-) -> Result<()> {
-    use std::time::{Duration, SystemTime, UNIX_EPOCH};
-    let stale_deadline = SystemTime::now().duration_since(UNIX_EPOCH)?;
-    let stale_deadline = stale_deadline - Duration::from_secs(48 * 60 * 60);
-    let stale_deadline = stale_deadline.as_millis() as i64;
-
-    let list = client
-        .list_datasets()
-        .set_project_id(project_id)
-        .set_filter(format!("labels.{INSTANCE_LABEL}"))
-        .by_item()
-        .into_stream();
-    let datasets = list.collect::<Vec<_>>().await;
-
-    let pending_all_datasets = datasets
-        .iter()
-        .filter_map(|v| match v {
-            Ok(v) => {
-                if let Some(dataset_id) = extract_dataset_id(project_id, &v.id) {
-                    return Some(
-                        client
-                            .get_dataset()
-                            .set_project_id(project_id)
-                            .set_dataset_id(dataset_id)
-                            .send(),
-                    );
-                }
-                None
-            }
-            Err(_) => None,
-        })
-        .collect::<Vec<_>>();
-
-    let stale_datasets = futures::future::join_all(pending_all_datasets)
-        .await
-        .into_iter()
-        .filter_map(|r| match r {
-            Ok(dataset) => Some(dataset),
-            Err(e) if e.status().is_some_and(|s| s.code == Code::NotFound) => None,
-            Err(_) => panic!("expected a successful get_dataset()"),
-        })
-        .filter_map(|dataset| {
-            if dataset
-                .labels
-                .get(INSTANCE_LABEL)
-                .is_some_and(|v| v == "true")
-                && dataset.creation_time < stale_deadline
-            {
-                return Some(dataset);
-            }
-            None
-        })
-        .collect::<Vec<_>>();
-
-    println!("found {} stale datasets", stale_datasets.len());
-
-    let pending_deletion: Vec<_> = stale_datasets
-        .into_iter()
-        .filter_map(|ds| {
-            if let Some(dataset_id) = extract_dataset_id(project_id, &ds.id) {
-                return Some(
-                    client
-                        .delete_dataset()
-                        .set_project_id(project_id)
-                        .set_dataset_id(dataset_id)
-                        .set_delete_contents(true)
-                        .send(),
-                );
-            }
-            None
-        })
-        .collect();
-
-    futures::future::join_all(pending_deletion).await;
-
-    Ok(())
-}
-
-pub(crate) fn random_dataset_id() -> String {
-    let rand_suffix = random_id_suffix();
-    format!("rust_bq_test_dataset_{rand_suffix}")
-}
-
-fn extract_dataset_id(project_id: &str, id: &str) -> Option<String> {
-    id.strip_prefix(project_id)
-        .and_then(|s| s.strip_prefix(":"))
-        .map(|v| v.to_string())
 }
