+++ tests/bigquery/src/lib.rs
@@ -16,12 +16,7 @@ mod dataset;
 mod job;
 mod query;
 mod reads;
-mod table;
 mod writes;
-
-use rand::{RngExt, distr::Alphanumeric};
-
-const INSTANCE_LABEL: &str = "rust-sdk-integration-test";
 
 pub use dataset::dataset_admin;
 pub use job::{job_service, job_service_poller, job_service_poller_error};
@@ -32,10 +27,4 @@ pub use query::{
 pub use reads::run_reads;
 pub use writes::run_writes;
 
-fn random_id_suffix() -> String {
-    rand::rng()
-        .sample_iter(&Alphanumeric)
-        .take(8)
-        .map(char::from)
-        .collect()
-}
+pub use bigquery_samples::{INSTANCE_LABEL, random_id_suffix};
