+++ tests/bigquery/src/writes.rs
@@ -14,12 +14,14 @@
 
 mod arrow;
 
-use crate::dataset::{cleanup_stale_datasets, create_dataset, delete_dataset, random_dataset_id};
 use anyhow::Result;
+use bigquery_samples::{
+    cleanup_stale_datasets, create_dataset, create_table, delete_dataset, random_dataset_id,
+};
 use google_cloud_bigquery::client::{BigQuery, Write};
 use google_cloud_bigquery::query::FromRow;
 use google_cloud_bigquery_v2::client::{DatasetService, TableService};
-use google_cloud_bigquery_v2::model::{Table, TableFieldSchema, TableReference, TableSchema};
+use google_cloud_bigquery_v2::model::{TableFieldSchema, TableSchema};
 use google_cloud_test_utils::runtime_config::project_id;
 
 pub async fn run_writes() -> Result<()> {
@@ -34,7 +36,12 @@ pub async fn run_writes() -> Result<()> {
     let table_id = "writes";
 
     let result = async {
-        create_writes_table(&table_service, &project_id, &dataset_id, table_id).await?;
+        let schema = TableSchema::new().set_fields([
+            TableFieldSchema::new().set_name("name").set_type("STRING"),
+            TableFieldSchema::new().set_name("age").set_type("INTEGER"),
+            TableFieldSchema::new().set_name("test").set_type("STRING"),
+        ]);
+        create_table(&table_service, &project_id, &dataset_id, table_id, schema).await?;
         let client = Write::builder().build().await?;
         arrow::basic(&client, &project_id, &dataset_id, table_id).await?;
         arrow::pending(&client, &project_id, &dataset_id, table_id).await?;
@@ -67,7 +74,7 @@ pub(crate) async fn read_writes_table(
     let mut rows = client
         .query(query)
         .with_project_id(project_id)
-        .set_labels(vec![(crate::INSTANCE_LABEL, "true")])
+        .set_labels(vec![(bigquery_samples::INSTANCE_LABEL, "true")])
         .until_done()
         .await?
         .read();
@@ -77,36 +84,4 @@ pub(crate) async fn read_writes_table(
         users.push(row?.try_into()?);
     }
     Ok(users)
-}
-
-pub(crate) async fn create_writes_table(
-    table_service: &TableService,
-    project_id: &str,
-    dataset_id: &str,
-    table_id: &str,
-) -> anyhow::Result<()> {
-    let schema = TableSchema::new().set_fields([
-        TableFieldSchema::new().set_name("name").set_type("STRING"),
-        TableFieldSchema::new().set_name("age").set_type("INTEGER"),
-        TableFieldSchema::new().set_name("test").set_type("STRING"),
-    ]);
-
-    table_service
-        .insert_table()
-        .set_project_id(project_id)
-        .set_dataset_id(dataset_id)
-        .set_table(
-            Table::new()
-                .set_table_reference(
-                    TableReference::new()
-                        .set_project_id(project_id)
-                        .set_dataset_id(dataset_id)
-                        .set_table_id(table_id),
-                )
-                .set_schema(schema),
-        )
-        .send()
-        .await?;
-
-    Ok(())
 }
